Add rotation to pyi
Trying having a better structure but it didn't work
